+++ src/scenes/Main.js
@@ -13,12 +13,16 @@ export default class Main extends Scene {
     this.load.image('star', 'assets/star.png');
     this.load.image('bomb', 'assets/bomb.png');
     this.load.spritesheet('player', 'assets/dude.png', { frameWidth: 32, frameHeight: 48 });
+
+    if (!window.localStorage.getItem('name')) {
+      window.localStorage.setItem('name', window.prompt('Please provide a name: '));
+    }
   }
 
   create() {
     this.score = 0;
     this.scoreText = this.add.text(16, 16, 'score: 0', { fontSize: '32px', fill: '#fff' });
-    this.gameOverText = this.add.text(320, 250, 'Game Over (Press any key)', { fontSize: '48px', fill: '#fff' }).setVisible(false).setDepth(2);
+    this.gameOverText = this.add.text(600, 250, 'Game Over (Press any key)', { fontSize: '48px', fill: '#fff' }).setVisible(false).setDepth(2).setOrigin(0.5);
 
     this.bg = this.add.tileSprite(640, 374, 1280, 768, 'sky').setDepth(-1);
 
@@ -55,7 +59,7 @@ export default class Main extends Scene {
     this.input.keyboard.on('keydown', () => {
       if (this.gameOver) {
         this.gameOver = false;
-        this.scene.start('Main');
+        this.scene.start('LeaderBoard');
       }
     });
   }