+++ CHANGELOG.md
@@ -15,3 +15,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Stars and bombs bouncing.
 - Score counter.
 - Restart the game after game over.
+- Prmop users for their name.
+- Leader board initial scene.